pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["vikyw89 <vikyw89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pyautogen = "^0.2.27"
python-dotenv = "^1.0.1"
yfinance = "^0.2.38"
matplotlib = "^3.8.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "scripts.start:run"